requirements.txt
django==5.2.6
djangorestframework
tweepy
python-social-auth
Pillow
facebook-business
django-cors-headers
djangorestframework-simplejwt
OAuth2
django-ratelimit
pymemcache